pyproject.toml
[tool.poetry]
name = "monicaapi"
version = "0.1.0"
description = ""
authors = ["Hongpei Zheng <zhenghongpei@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
